refactor(mocks): extract cliTokens module from mockApi (#991 M3)
Move CLI-token state/helpers (cliTokens, cliAuthRequests, CROCKFORD32, normalizeMockUserCode, stripOwner) and the 7 listCliTokens…denyCliAuth methods into cliTokens.ts as cliTokensApi, spread into index.ts.
